Modification des places pour ConfigurationSecurityApplication, CustomUserDetail vers security.
ConfigurationSecurityApplication implémente désormais CORS

nouveau package jwt qui inclut JwtFilter

package configuration qui va implémenter SwaggerConfig et WebConfig
